Automatically backed up by Learn

+++ lib/bachelor.rb
@@ -54,7 +54,7 @@ def get_average_age_for_season(data, season)
   data.each do |season_number, contestant_array|
     if season_number == season 
       contestant_array.each do |contestant_hash|
-        age_total += contestant_hash["age"].to_i
+        age_total += contestant_hash["age"].to_f 
         num_of_contestants += 1 
       end
     end
